Change groupid org.hawkular.client to org.hawkular.metrics
