+++ hawkular-dropwizard-reporter-factory/src/main/resources/META-INF/services/io.dropwizard.metrics.ReporterFactory
@@ -1,1 +1,1 @@
-org.hawkular.client.dropwizard.HawkularReporterFactory
+org.hawkular.metrics.dropwizard.HawkularReporterFactory
