+++ hawkular-dropwizard-reporter/src/test/java/org/hawkular/metrics/dropwizard/HawkularReporterTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.client.dropwizard;
+package org.hawkular.metrics.dropwizard;
 
 import static java.util.stream.Collectors.toMap;
 
@@ -30,8 +30,8 @@ import java.util.stream.StreamSupport;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.assertj.core.api.iterable.Extractor;
-import org.hawkular.client.http.HawkularHttpClient;
-import org.hawkular.client.http.HawkularHttpResponse;
+import org.hawkular.metrics.reporter.http.HawkularHttpClient;
+import org.hawkular.metrics.reporter.http.HawkularHttpResponse;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.Test;
@@ -128,6 +128,8 @@ public class HawkularReporterTest {
         assertThat(client.getPostedTags()).containsOnly(
                 Pair.of("/counters/my.histogram.count/tags", "{\"histogram\":\"count\"}"),
                 Pair.of("/gauges/my.histogram.mean/tags", "{\"histogram\":\"mean\"}"),
+                Pair.of("/gauges/my.histogram.min/tags", "{\"histogram\":\"min\"}"),
+                Pair.of("/gauges/my.histogram.max/tags", "{\"histogram\":\"max\"}"),
                 Pair.of("/gauges/my.histogram.stddev/tags", "{\"histogram\":\"stddev\"}"),
                 Pair.of("/gauges/my.histogram.median/tags", "{\"histogram\":\"median\"}"),
                 Pair.of("/gauges/my.histogram.75perc/tags", "{\"histogram\":\"75perc\"}"),
