+++ hawkular-reporter-common/src/main/java/org/hawkular/metrics/reporter/http/HawkularJson.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.client.json;
+package org.hawkular.metrics.reporter.http;
 
 import java.math.BigDecimal;
