+++ hawkular-reporter-common/src/main/java/org/hawkular/metrics/reporter/http/JdkHawkularHttpClient.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.client.http;
+package org.hawkular.metrics.reporter.http;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
