+++ pom.xml
@@ -26,7 +26,7 @@
     <version>43</version>
   </parent>
 
-  <groupId>org.hawkular.client</groupId>
+  <groupId>org.hawkular.metrics</groupId>
   <artifactId>hawkular-dropwizard-reporter-parent</artifactId>
   <version>0.1.0-SNAPSHOT</version>
   <packaging>pom</packaging>
@@ -76,12 +76,12 @@
   <dependencyManagement>
     <dependencies>
       <dependency>
-        <groupId>org.hawkular.client</groupId>
+        <groupId>org.hawkular.metrics</groupId>
         <artifactId>hawkular-reporter-common</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
-        <groupId>org.hawkular.client</groupId>
+        <groupId>org.hawkular.metrics</groupId>
         <artifactId>hawkular-dropwizard-reporter</artifactId>
         <version>${project.version}</version>
       </dependency>
